external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/embedded-struct-definitions-forbidden-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:3: 'Embedded struct definitions are not allowed' :  

embedded structure definitions are forbidden per GLSL ES section 4.1.8, "Structures", and should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: embedded structure definitions are forbidden per GLSL ES section 4.1.8, "Structures", and should fail
test vertex shader
default fragment shaderPASS embedded structure definitions are forbidden per GLSL ES section 4.1.8, "Structures", and should fail
PASS successfullyParsed is true

TEST COMPLETE
